Reregister app
