+++ apps/apps.go
@@ -129,6 +129,8 @@ import (
 	"tidbyt.dev/community/apps/nhlstandings"
 	"tidbyt.dev/community/apps/nightscout"
 	"tidbyt.dev/community/apps/nixelclock"
+	"tidbyt.dev/community/apps/njtransit"
+	"tidbyt.dev/community/apps/njtransitdepartures"
 	"tidbyt.dev/community/apps/noaabuoy"
 	"tidbyt.dev/community/apps/noaatides"
 	"tidbyt.dev/community/apps/nyancat"
@@ -350,6 +352,8 @@ func GetManifests() []manifest.Manifest {
 		nhlstandings.New(),
 		nightscout.New(),
 		nixelclock.New(),
+		njtransit.New(),
+		njtransitdepartures.New(),
 		noaabuoy.New(),
 		noaatides.New(),
 		nyancat.New(),
