+++ apps/njtransit/README.md
@@ -1,0 +1,4 @@
+# NJ Transit Depature Vision viewer
+
+Displays the trains and their departing times for a given NJ Transit station.
+Station List is a hardcoded CSV but shouldnt need to be updated often.
